refactor: user 패키지명을 member로 변경
