+++ src/main/java/antigravity/application/ProductService.java
@@ -5,8 +5,8 @@ import antigravity.application.exception.ProductNotFoundException;
 import antigravity.application.exception.UserNotFoundException;
 import antigravity.domain.product.Product;
 import antigravity.domain.product.ProductRepository;
-import antigravity.domain.user.Member;
-import antigravity.domain.user.MemberRepository;
+import antigravity.domain.member.Member;
+import antigravity.domain.member.MemberRepository;
 import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
