+++ src/main/java/antigravity/infra/respository/JpaMemberRepository.java
@@ -1,7 +1,7 @@
 package antigravity.infra.respository;
 
-import antigravity.domain.user.Member;
-import antigravity.domain.user.MemberRepository;
+import antigravity.domain.member.Member;
+import antigravity.domain.member.MemberRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaMemberRepository extends MemberRepository, JpaRepository<Member, Long> {
